requirements.txt
customtkinter
python-bidi
requests
